requirements.txt
# AgriSmart Backend Dependencies
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Machine Learning dependencies
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
joblib==1.3.2

# TensorFlow for .h5 model support (optional)
tensorflow==2.15.0

# Additional ML libraries
scipy==1.11.4

# HTTP client
httpx==0.25.2

# Validation
pydantic==2.5.0
email-validator==2.1.0

# Logging and monitoring
structlog==23.2.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0